requirements.txt
Django==1.10.4
djangorestframework==3.5.3
pymongo==3.4.0
nsapi==2.7.4